531-django-series/531b-mmcard/requirements.txt

django>=2.2,<3.0
python-dotenv==0.10.3
# psycopg2==2.8.3
gunicorn==19.9.0
mysqlclient==1.4.6
django-mysql<2.5
django-bootstrap3>12.0
#bootstrap-admin
django-crudbuilder
#django-tables2
djangorestframework
